requirements.txt
aiogram
django
psycopg2-binary
python-dotenv